+++ src/brod_group_subscriber_worker.erl
@@ -21,8 +21,8 @@
 
 -include("brod_int.hrl").
 
-%% Brod callbacks
--export([init/2, handle_message/3]).
+%% brod_topic_subscriber callbacks
+-export([init/2, handle_message/3, terminate/2]).
 
 -type start_options() ::
         #{ group_id     := brod:group_id()
@@ -97,6 +97,9 @@ handle_message(_Partition, Msg, State) ->
       NewState = State#state{cb_state = NewCbState},
       {ok, NewState}
   end.
+
+terminate(Reason, #state{cb_module = CbModule, cb_state = State}) ->
+  brod_utils:optional_callback(CbModule, terminate, [Reason, State], ok).
 
 %%%===================================================================
 %%% Internal functions