+++ test/brod_test_group_subscriber.erl
@@ -27,6 +27,7 @@
         , get_committed_offset/3
         , handle_message/2
         , assign_partitions/3
+        , terminate/2
         ]).
 
 init(InitInfo, Config) ->
@@ -74,3 +75,12 @@ assign_partitions(_CbConfig, Members, TopicPartitions) ->
   PartitionsAssignments = [{Topic, [PartitionsN]}
                            || {Topic, PartitionsN} <- TopicPartitions],
   [{element(1, hd(Members)), PartitionsAssignments}].
+
+terminate(Reason, #state{ topic = Topic
+                        , partition = Partition
+                        }) ->
+  ?tp(brod_test_group_subscriber_terminate,
+      #{ topic     => Topic
+       , partition => Partition
+       , readon    => Reason
+       }).